third_party/blink/web_tests/platform/generic/media/W3C/video/networkState/networkState_during_progress-expected.txt

videoElement.networkState should be NETWORK_LOADING during progress event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL "1" should be 2. Was 1.
PASS successfullyParsed is true

TEST COMPLETE


spec reference
